pyproject.toml
[project]
name = "jobrecom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.0.15",
    "docx2txt>=0.9",
    "faiss-cpu>=1.11.0.post1",
    "groq>=0.31.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.72",
    "langchain-groq>=0.3.7",
    "pdfplumber>=0.11.7",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.0",
    "streamlit>=1.48.0",
]
